[LS-9] 고가용성설계 - circuit breaker, kakao-client 적용

+++ external-api/kakao-client/src/main/java/com/library/Meta.java
@@ -8,6 +8,6 @@ public record Meta(
         @JsonProperty("pageable_count")
         int pageableCount,
         @JsonProperty("total_count")
-        long totalCount
+        int totalCount
 ) {
 }
